requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
Pillow>=10.0.0
requests>=2.31.0
Werkzeug>=2.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
google-auth==2.25.2
google-auth-oauthlib==1.2.0
